+++ resources/ReleaseNotes.md
@@ -3,3 +3,4 @@
 - Allow cropping of `mag.Science` with negative duration (filters from the end)
 - Allow cropping of I-ALiRT with separate filter for primary and secondary
 - Add conversion method `eventtable` in `mag.event.Event` to simplify logic in science analysis
+- Add tests for `mag.event.Event`, `mag.event.ModeChange`, `mag.event.RangeChange` and `mag.event.RampMode`